Tpetra: Protect all deprecated things with enable macro
@trilinos/tpetra

For everything in Tpetra that is marked deprecated with the
TPETRA_DEPRECATED macro, wrap it with the
TPETRA_ENABLE_DEPRECATED_CODE macro, so that the code no longer exists
if that macro is not defined.

I took the opportunity to clean up some other declarations and improve
Tpetra's Doxygen documentation.